C++ source compatibility.
In C, `this` is not a reserved keyword, but it is in C++; this rev
removes it where used.

Fixes an undercover bug by which an object was being created with an
wrong, invalid size.

Makes a provision for a configurable `*.h` or `*.hpp` extension, even
not using it for now (currently using `*.h` for either C or C++
builds.

Added multiple missing assertions absent after `malloc` calls checking
that `NULL` pointers aren't received.

Able to build using either `clang` or `clang++`, guaranteeing parallel
build support for up to 4 concurrent `make` _jobs_, for either `clang`
or `clang++`.

CI will test this using GCC `4.8`.